modification ketkout

+++ src/main/java/com/example/backofficepro/model/Category.java
@@ -16,7 +16,7 @@ public class Category {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     private String name;
     private String description;
@@ -24,7 +24,7 @@ public class Category {
     @OneToMany(mappedBy = "category")  // Le mappedBy fait référence à 'category' dans Media (TVShow, TVMovie)
     private List<Media> mediaList;  // Changez ici en List<Media>
 
-    public Category(Integer id, String name, String description) {
+    public Category(Long id, String name, String description) {
         this.id = id;
         this.name = name;
         this.description = description;
