+++ src/main/java/com/example/backofficepro/model/TVMovie.java
@@ -18,13 +18,11 @@ public class TVMovie extends Media {
     private String movieUrl;
     private String part;  // Partie spécifique au film (ex: Partie 1 de Harry Potter)
 
-    public TVMovie(Integer id, String title, Integer rating, String description, String photoUrl, String releaseDate, Integer duration, String trailerUrl, String movieUrl, String part) {
+    public TVMovie(Long id, String title, Integer rating, String description, String photoUrl, String releaseDate, Integer duration, String trailerUrl, String movieUrl, String part) {
         super(id, title, rating, description, photoUrl, releaseDate);
         this.duration = duration;
         this.trailerUrl = trailerUrl;
         this.movieUrl = movieUrl;
         this.part = part;
     }
-
-    // Getters and setters omitted for brevity
 }
