+++ src/main/java/com/example/backofficepro/orchestration/ThemeOrchestration.java
@@ -7,13 +7,22 @@ import org.springframework.stereotype.Component;
 
 import java.util.List;
 
+/**
+ * Orchestrateur du référentiel thématique.
+ * Facilite :
+ * - La gestion de la taxonomie des thèmes
+ * - Les opérations de recherche avancée
+ * - L'intégration avec les systèmes de tagging
+ */
+
+
 @Component
 public class ThemeOrchestration {
 
     @Autowired
     private IThemeService themeService;
 
-    public ThemeDTO getThemeById(Integer id) {
+    public ThemeDTO getThemeById(Long id) {
         return themeService.getThemeById(id);
     }
 
@@ -25,11 +34,11 @@ public class ThemeOrchestration {
         return themeService.createTheme(themeDTO);
     }
 
-    public ThemeDTO updateTheme(Integer id, ThemeDTO themeDTO) {
+    public ThemeDTO updateTheme(Long id, ThemeDTO themeDTO) {
         return themeService.updateTheme(id, themeDTO);
     }
 
-    public void deleteTheme(Integer id) {
+    public void deleteTheme(Long id) {
         themeService.deleteTheme(id);
     }
 }
