+++ src/main/java/com/example/backofficepro/repository/ThemeRepository.java
@@ -7,6 +7,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 
-public interface ThemeRepository extends JpaRepository<Theme, Integer> {
+public interface ThemeRepository extends JpaRepository<Theme, Long> {
     //List<Theme> findByNameContainingIgnoreCase(String name);
 }
